Intent: Strengthen graph metadata access coherence by centralising helpers. Operators involved: [Coupling, Resonance] Affected invariants: [#1, #4, #7] Key changes: - Move get_graph and get_graph_mapping into tnfr.graph_utils for canonical access. - Update helper modules and consumers to use the centralised graph utilities. - Preserve helper exports and mark_dnfr_prep_dirty behaviour after relocation. Expected risks/dissonances: Minor import regressions; mitigated by re-exporting functions. Metrics: C(t) stable; νf unaffected; phase unchanged. Equivalence map: get_graph -> tnfr.graph_utils.get_graph; get_graph_mapping -> tnfr.graph_utils.get_graph_mapping.
